+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [0.0.5] / 2022-01-13
+- Add Move RibbonPanelTo
+- Add Remove RibbonTab
+- Add `GetRibbonPanel` using `m_RibbonPanel`
+- Remove Panel
+- Add OrderPanels
 - Clear Code
 - Change to Now.Tick
 - Add ConsoleAttribute